World Indicator Analysis Using Tableau
This dataset is taken from tableau itself. And this data is provided by World Bank. It is about various development indicators for the world. This data includes 207 countries 6 regions and has records like ease of business ranks, business tax rate, birth rate, gdp etc. that we will know when I explain the dashboards further.
This dataset can be useful for government policy makers, economists, investors, businessman, educational institutions. But I have made dashboards from the business perspective. So I would like you to perceive this dashboard as a businessman who wants to expand or start a business in other country or countries and is looking at the economy and other aspects of various countries.